Added routing + moved Login logic from App.js to Home.js
