+++ react-login/src/components/NavBar.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function NavBar() {
+    return (
+        <div className="Navbar">
+            HELLO THIS IS THE NAVBAR
+        </div>
+    );
+}
+
+export default NavBar;
