+++ react-login/src/pages/Receptionist.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Receptionist() {
+    return (
+        <div className="Receptionist">
+            HELLO THIS IS THE RECEPTIONIST
+        </div>
+    );
+}
+
+export default Receptionist;
